See 445485b1

+++ app/common/data/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Integer, LargeBinary, String
+from sqlalchemy import Boolean, Column, DateTime, ForeignKey, BigInteger, LargeBinary, String
 from sqlalchemy.orm import relationship
 
 from datetime import datetime
@@ -10,7 +10,7 @@ class BaseEntity(Base):
 
     __abstract__ = True
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     created_on = Column(DateTime, default=datetime.utcnow, nullable=False)
     updated_on = Column(DateTime, default=None, onupdate=datetime.utcnow, nullable=True)
     is_deleted = Column(Boolean, default=False, nullable=False)
@@ -38,6 +38,6 @@ class UserToken(BaseEntity):
 
     token = Column(String, nullable=False, index=True)
     token_type = Column(String, nullable=False)
-    expiry = Column(Integer, nullable=False)
-    user_id = Column(Integer, ForeignKey("users.id"))
+    expiry = Column(BigInteger, nullable=False)
+    user_id = Column(BigInteger, ForeignKey("users.id"))
     user = relationship("User")
